przeniesienie funkcji wyszukiwania filmów do search-movies.js
