+++ src/index.js
@@ -1,4 +1,4 @@
 import './sass/main.scss';
-
+import './js/search-movies';
 import './js/fetch-data';
 import './js/modal.js';
