+++ src/js/fetch-data.js
@@ -1,12 +1,8 @@
 const galleryOfMovies = document.querySelector('.movie-gallery');
-const input = document.querySelector('.search-form__input');
-const searchBtn = document.querySelector('.search-form__button');
-
 const API_KEY = '50faffa66bb05e881b7f3de0b265b30c';
 const BASE_URL = 'https://api.themoviedb.org/3';
 const MAIN_PAGE_PATH = '/trending/all/day';
 const GENRE_LIST_PATH = `/genre/movie/list`;
-const SEARCH_MOVIE_PATH = `/search/movie`;
 
 let page = 1;
 
@@ -54,26 +50,5 @@ async function fetchGenres() {
 fetchMovies(page)
   .then(renderMoviesCards)
   .catch(error => console.error(error));
-function clearInterfaceUI() {
-  galleryOfMovies.innerHTML = '';
-}
 
-searchBtn.addEventListener('click', searchMovies);
-
-async function searchMovies(e) {
-  e.preventDefault();
-  const response = await fetch(
-    `${BASE_URL}${SEARCH_MOVIE_PATH}?api_key=${API_KEY}&query=${input.value}&page=${page}`,
-  );
-  const data = await response.json();
-  fetchMovies(input.value)
-    .then(renderMoviesCards(data.results))
-    .catch(error => {
-      clearInterfaceUI();
-      alert('No results');
-    })
-    .finally((galleryOfMovies.innerHTML = ''));
-  return (galleryOfMovies.innerHTML = data.query[0]);
-}
-
-export { fetchMovies, renderMoviesCards, fetchGenres, searchMovies };
+export { fetchMovies, renderMoviesCards, fetchGenres };
